% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindClusters.R
\name{FindClusters}
\alias{FindClusters}
\title{FindClusters}
\usage{
FindClusters(df)
}
\arguments{
\item{df}{A data frame containing columns 'bitSeq' with sequences to be clustered and 'bitID' with sequence names.}
}
\value{
An object of class 'hclust' representing the hierarchical clustering of sequences.
}
\description{
This function performs hierarchical clustering on sequences based on sequence alignment scores.
}
\examples{
# clustering <- FindClusters(Hitsframe)

}
