% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateHitEnrichment.R
\name{FindEnrichedMotifs}
\alias{FindEnrichedMotifs}
\title{Find conserved sequences that are enriched specifically in your promoters of interest vs. control promoters}
\usage{
FindEnrichedMotifs(
  Job_name,
  comp_filter = TRUE,
  com_threshold = 0.9,
  bit_size = 16,
  bit_step = 5,
  promoter_size = 1000,
  min_bitratio = 1.5,
  num_threads = 3,
  min_match_size = 4,
  gap_open = 0,
  gap_extend = 2,
  penalty_mismatch = -2,
  match_reward = 1,
  e_value = 10,
  promoters_input = "input.txt",
  c_promoters_input = "control.txt"
)
}
\arguments{
\item{promoters_input}{Promoters of interest input file}

\item{c_promoters_input}{Control promoters input file}
}
\value{
A hitsframe containing enrichment of each specific bit
}
\description{
This function takes in an input list of promoters of interest, control promoters, and a custom database of promoters.
It splits up promoters of interest into "bits" and BLASTs each bit against all other promoters.
Promoters that are found more in your promoters of interest list, than in the control promoters list are considered enriched.
Several parameters can be ajusted to increase or decrease precision of BLASTs, and to reduce running time.
}
