% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWMFromHits.R
\name{PWMFromHits}
\alias{PWMFromHits}
\title{This function returns PWMs and seqlogos of enriched queries.}
\usage{
PWMFromHits(
  min_bitratio = 1.5,
  num_threads = 3,
  min_match_size = 4,
  gap_open = 0,
  gap_extend = 2,
  penalty_mismatch = -2,
  match_reward = 1,
  e_value = 10,
  promoter_data = "All Tomato 2k promoters.csv",
  input = input
)
}
\arguments{
\item{input}{A hitsframe containing enriched queries}
}
\value{
A results dataframe containing PWMs and plots for each input query
}
\description{
This function returns PWMs and seqlogos of enriched queries.
}
