% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowEnrichedLocations.R
\name{ShowEnrichedLocations}
\alias{ShowEnrichedLocations}
\title{Align and highlight enriched bits on promoters
Be sure to have performed a thorough preselection before using this function}
\usage{
ShowEnrichedLocations(
  Job_name = "Test-y",
  min_bitratio = 1.5,
  num_threads = 3,
  min_match_size = 4,
  gap_open = 0,
  gap_extend = 2,
  penalty_mismatch = -2,
  match_reward = 1,
  e_value = 10,
  promoter_data = "All Tomato 2k promoters.csv",
  data = Hitsframe,
  promoter_size = 100,
  Comp_Filter = TRUE,
  filter_strength = 0.8,
  return_f = "txt"
)
}
\description{
Align and highlight enriched bits on promoters
Be sure to have performed a thorough preselection before using this function
}
